requirements.txt
-i https://pypi.org/simple
alembic==1.8.1
-e git+https://github.com/dunkmann00/PyAPNs2.git@4622972650941a05fe6d4a58fb2e403342133aa9#egg=apns2
bcrypt==4.0.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
dnspython==2.2.1
email-validator==1.2.1
flask==2.2.2
flask-httpauth==4.7.0
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
flask-talisman==1.0.0
flask-wtf[email]==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
jinja2==3.1.2
lxml==4.9.1
mako==1.2.2
markupsafe==2.1.1
passlib[bcrypt]==1.7.4
phonenumbers==8.12.55
-e git+https://github.com/dunkmann00/premailer.git@08ea63266ed9342e0f3acf381bfe731b96dbe035#egg=premailer
psycopg2==2.9.3
pycparser==2.21
pyjwt==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
setuptools==59.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlalchemy[postgresql]==1.4.41
tomlkit==0.11.4
twilio==7.14.0
urllib3==1.26.12
werkzeug==2.2.2
wtforms==3.0.1
zipp==3.8.1
